refactoring and auth. channel selection
