+++ src/main/java/ee/ria/sso/EchoingPrincipalResolver.java
@@ -1,5 +1,6 @@
 package ee.ria.sso;
 
+
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.builder.ToStringBuilder;
 import org.apereo.cas.authentication.AuthenticationHandler;
@@ -10,29 +11,30 @@ import org.apereo.services.persondir.IPersonAttributeDao;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 
+
 /**
  * Created by serkp on 14.09.2017.
  */
 @Component
 @Qualifier(value = "echoingPrincipalResolver")
 public class EchoingPrincipalResolver implements PrincipalResolver {
-	public EchoingPrincipalResolver() {
-	}
+    public EchoingPrincipalResolver() {
+    }
 
-	public Principal resolve(Credential credential, Principal principal,
-			AuthenticationHandler handler) {
-		return principal;
-	}
+    public Principal resolve(Credential credential, Principal principal,
+                             AuthenticationHandler handler) {
+        return principal;
+    }
 
-	public boolean supports(Credential credential) {
-		return StringUtils.isNotBlank(credential.getId());
-	}
+    public boolean supports(Credential credential) {
+        return StringUtils.isNotBlank(credential.getId());
+    }
 
-	public String toString() {
-		return (new ToStringBuilder(this)).toString();
-	}
+    public String toString() {
+        return (new ToStringBuilder(this)).toString();
+    }
 
-	public IPersonAttributeDao getAttributeRepository() {
-		return null;
-	}
+    public IPersonAttributeDao getAttributeRepository() {
+        return null;
+    }
 }
