+++ src/main/java/ee/ria/sso/MIDAuthenticator.java
@@ -1,10 +1,11 @@
 package ee.ria.sso;
 
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.stereotype.Component;
 
 import com.codeborne.security.mobileid.MobileIDAuthenticator;
 import com.codeborne.security.mobileid.MobileIDSession;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.stereotype.Component;
+
 
 /**
  * Created by serkp on 21.09.2017.
@@ -13,9 +14,9 @@ import com.codeborne.security.mobileid.MobileIDSession;
 @Qualifier(value = "MIDAuthenticator")
 public class MIDAuthenticator extends MobileIDAuthenticator {
 
-	@Override
-	protected MobileIDSession startLogin(String personalCode, String countryCode, String phone) {
-		return super.startLogin(personalCode, countryCode, phone);
-	}
+    @Override
+    protected MobileIDSession startLogin(String personalCode, String countryCode, String phone) {
+        return super.startLogin(personalCode, countryCode, phone);
+    }
 
 }
