+++ src/main/webapp/WEB-INF/classes/messages.properties
@@ -14,3 +14,4 @@ ria.mid.error.text=Autentimine ei õnnestunud.
 ria.mid.error.servicelimited.title=Teenusele ligipääs piiratud!
 label.idcard.title=Sisenege ID-Kaardiga
 label.idcard.instruction=Sisestage oma ID-Kaart kaardilugejasse ja vajutage Jätkan nuppu.
+ria.back.to.auth=Vajutage siia, et minna tagasi autentimiskanali valikusse!
